restructuring towards julia project
created a TOFTracer2 main folder and a /src folder within that contains the code.
Also, the startup.jl is renamed to TOFTracer2.jl because this is needed
for the project structure and TOFTracer2 became a module by itself.
It is now exporting the regularly needed modules to keep the code short
(no need to add TOFTracer2. everywhere)
Added Manifest and Project File.
