+++ README.md
@@ -1,5 +1,7 @@
 # Julia lang data processing scripts for Tofwerk based PTR TOF Data
 
+
+
 This software provides functions to extract time traces from raw TOF data based on mass border integration and peakshape deconvolution from predefined mass lists including substraction of known isotopes, mass scale correction and automatic peakshape calculation.
 
 ## Prerequisites
@@ -23,7 +25,7 @@ open VS Code, add julia extension
 
 ## Getting Started:
 
-start julia from base directory ("TOF-Tracer2/")
+start julia from base directory ("TOFTracer2/")
 add packages ( press "]" to enter Pkg mode, run "add PyPlot HDF5 DSP", leave Pkg mode with backspace)
 Try the example with include("processingProjects/processingProject-example.jl") -> change all paths to the paths on your PC,
