+++ src/TOFTracer2.jl
@@ -1,14 +1,20 @@
+module TOFTracer2
+
 using Distributed
 using Dates
 using DelimitedFiles
 using PyPlot
+
+export correctMassScaleAndExtractSumSpec, baselineAndPeakshape, deconvolute, ResultFileFunctions, MasslistFunctions, PlotFunctions, InterpolationFunctions, massLibrary
 
 if (nprocs() < 2 && isdefined(Main, :usePrecaching) && usePrecaching)
     println("Adding process for file precaching!")
     addprocs(1)
 end
 
-scriptspath =  pwd()
+scriptspath = @__DIR__
+print("scriptspath: ",scriptspath, "\n")
+
 if ! isdefined(Main, :FunctionsLoaded)
     #load includes
     include("includes/BaselineFunctions.jl")
@@ -23,11 +29,13 @@ if ! isdefined(Main, :FunctionsLoaded)
     include("includes/TofTracer.jl")
 
     #Load Personal MassLibrary Definitions
-    include("$(scriptspath)/manualMassLibrary.jl")
+    include("includes/manualMassLibrary.jl")
 
     #Load workflows
     include("$(scriptspath)/workflows/combinedMassScaleAndExtractSumSpec.jl")
     include("$(scriptspath)/workflows/peakShape.jl")
     include("$(scriptspath)/workflows/deconvolutionMatrix.jl")
     FunctionsLoaded = true
+end
+
 end
