+++ src/includes/MultipeakFunctionsAPi.jl
